server/pyproject.toml
[tool.poetry]
name = "vrr-server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
mss = "^7.0.1"
ffmpeg-python = "^0.2.0"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
vrr = "vrr_server:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
